ELEMENTS-1616: move thumbnail fallback helper into common-utils per review
elements/ is reserved for Polymer custom elements, but nuxeo-thumbnail-fallback.js
held only a shared constant plus a one-line helper. Per review r3718785608, move
BLANK_THUMBNAIL_SRC and applyThumbnailFallback into the existing shared utility
elements/common-utils.js, repoint the grid and list thumbnail imports to it, and
delete the now-empty nuxeo-thumbnail-fallback.js. Behaviour (crossorigin +
on-error transparent-pixel fallback) is unchanged.
