--- elements/nuxeo-thumbnail-fallback.js
@@ -1,30 +1,0 @@
-/**
-@license
-©2023 Hyland Software, Inc. and its affiliates. All rights reserved. 
-All Hyland product names are registered or unregistered trademarks of Hyland Software, Inc. or its affiliates.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-// ELEMENTS-1616: shared transparent 1x1 pixel used as the thumbnail fallback when a
-// (cross-origin) image request fails to load, so views degrade gracefully instead of
-// showing a broken-image icon. Kept in one place so the grid and list thumbnails stay
-// consistent with nuxeo-document-thumbnail.
-export const BLANK_THUMBNAIL_SRC =
-  'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAA' +
-  'C0lEQVR42mNkYAAAAAYAAjCB0C8AAAAASUVORK5CYII=';
-
-// Swaps a failed thumbnail <img> for the transparent pixel above.
-export function applyThumbnailFallback(img) {
-  img.src = BLANK_THUMBNAIL_SRC;
-}
